requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.12.3
cryptography==2.6.1
Django==2.2.1
django-cors-headers==3.0.1
django-filter==2.1.0
django-rest-knox==4.0.1
djangorestframework==3.9.4
drf-extensions==0.5.0
ecdsa==0.13.2
future==0.17.1
gunicorn==19.9.0
importlib-metadata==0.17
Markdown==3.1.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pyparsing==2.4.0
pytest==4.6.1
pytest-django==3.5.0
python-jose==3.0.1
pytz==2019.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
zipp==0.5.1
